refactor: Move controlers to root irectory
