+++ controllers/workspaceController.ts
@@ -1,4 +1,4 @@
-import connectMongo from '../lib/connectToMongo';
+import connectMongo from '../src/lib/connectToMongo';
 import { Types } from 'mongoose';
 import Workspace from 'schemas/Workspace';
 import User from 'schemas/User';
